+++ server.py
@@ -31,7 +31,7 @@ from utils import safe_filename
 from voice_designer import VoiceDesigner
 
 APP_NAME = "SoftMeta Chatterbox TTS Server"
-APP_VERSION = "0.3.0"
+APP_VERSION = "0.3.1"
 logger = logging.getLogger("softmeta.chatterbox")
 
 config = load_config()